requirements.txt
docopt==0.6.2
GitPython==0.3.2.RC1
